Replace booster by example
Related to: SB-1345
