+++ cute-name-service/.openshiftio/application.yaml
@@ -109,7 +109,7 @@ objects:
       app: spring-boot-cache-cutename
       provider: snowdrop
       version: "BOOSTER_VERSION"
-      group: io.openshift.booster
+      group: dev.snowdrop.example
     name: spring-boot-cache-cutename
   spec:
     ports:
@@ -120,7 +120,7 @@ objects:
     selector:
       app: spring-boot-cache-cutename
       provider: snowdrop
-      group: io.openshift.booster
+      group: dev.snowdrop.example
 - apiVersion: v1
   kind: DeploymentConfig
   metadata:
@@ -128,7 +128,7 @@ objects:
       app: spring-boot-cache-cutename
       provider: snowdrop
       version: "BOOSTER_VERSION"
-      group: io.openshift.booster
+      group: dev.snowdrop.example
     name: spring-boot-cache-cutename
   spec:
     replicas: 1
@@ -136,7 +136,7 @@ objects:
     selector:
       app: spring-boot-cache-cutename
       provider: snowdrop
-      group: io.openshift.booster
+      group: dev.snowdrop.example
     strategy:
       rollingParams:
         timeoutSeconds: 3600
@@ -147,7 +147,7 @@ objects:
           app: spring-boot-cache-cutename
           provider: snowdrop
           version: "BOOSTER_VERSION"
-          group: io.openshift.booster
+          group: dev.snowdrop.example
       spec:
         containers:
         - env:
@@ -199,7 +199,7 @@ objects:
       app: spring-boot-cache-cutename
       provider: snowdrop
       version: "BOOSTER_VERSION"
-      group: io.openshift.booster
+      group: dev.snowdrop.example
     name: spring-boot-cache-cutename
   spec:
     port:
