+++ cute-name-service/pom.xml
@@ -4,15 +4,15 @@
     <modelVersion>4.0.0</modelVersion>
 
     <parent>
-        <groupId>io.openshift.booster</groupId>
-        <artifactId>spring-boot-cache-parent</artifactId>
+        <groupId>dev.snowdrop.example</groupId>
+        <artifactId>cache</artifactId>
         <version>2.1.3-3-SNAPSHOT</version>
     </parent>
 
     <artifactId>spring-boot-cache-cutename</artifactId>
 
-    <name>Spring Boot - Cache Booster - Cute Name Service</name>
-    <description>Spring Boot - Cache Booster - Cute Name Service</description>
+    <name>Spring Boot - Cache Example - Cute Name Service</name>
+    <description>Spring Boot - Cache Example - Cute Name Service</description>
 
 
     <dependencies>
