+++ cute-name-service/src/main/java/dev/snowdrop/example/ExampleApplication.java
@@ -14,29 +14,20 @@
  * limitations under the License.
  */
 
-package io.openshift.booster;
+package dev.snowdrop.example;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cache.annotation.EnableCaching;
-import org.springframework.context.annotation.Bean;
-import org.springframework.web.client.RestTemplate;
 
 /**
  * Entry point to the application.
  */
 @SpringBootApplication
-@EnableCaching
-public class BoosterApplication {
+public class ExampleApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(BoosterApplication.class, args);
+        SpringApplication.run(ExampleApplication.class, args);
     }
 
-
-    @Bean
-    public RestTemplate restTemplate() {
-        return new RestTemplate();
-    }
 }
