+++ cute-name-service/src/main/java/dev/snowdrop/example/service/UserNameGenerator.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.openshift.booster.service;
+package dev.snowdrop.example.service;
 
 import java.util.HashSet;
 import java.util.Set;
