+++ greeting-service/src/main/java/dev/snowdrop/example/service/NameService.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.openshift.booster.service;
+package dev.snowdrop.example.service;
 
 public interface NameService {
