+++ greeting-service/src/main/resources/static/index.html
@@ -3,7 +3,7 @@
 <head>
 
     <meta charset="utf-8">
-    <title>Cache Mission - Spring Boot Booster</title>
+    <title>Cache Mission - Spring Boot Example</title>
 
     <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css"
           integrity="sha384-BVYiiSIFeK1dGmJRAkycuHAHRg32OmUcww7on3RYdg4Va+PmSTsz/K68vbdEjh4u" crossorigin="anonymous">
@@ -15,10 +15,10 @@
 <div class="container">
 
     <div>
-        <h1>Spring Boot Cache Booster</h1>
+        <h1>Spring Boot Cache Example</h1>
 
         <p class="lead">
-            The Caching booster demonstrates how to use a cache to increase the response time of applications.
+            The Caching example demonstrates how to use a cache to increase the response time of applications.
             It contains three components: the <code>greeting</code> service, the <code>cute name</code> service, and a remote cache.
         </p>
         <div class="lead">
