+++ pom.xml
@@ -1,12 +1,12 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
-  <artifactId>spring-boot-cache-parent</artifactId>
+  <artifactId>cache</artifactId>
   <version>2.1.3-3-SNAPSHOT</version>
   <packaging>pom</packaging>
-  <groupId>io.openshift.booster</groupId>
-  <name>Spring Boot - Cache Booster</name>
-  <description>Spring Boot - Cache Booster</description>
+  <groupId>dev.snowdrop.example</groupId>
+  <name>Spring Boot - Cache Example</name>
+  <description>Spring Boot - Cache Example</description>
   <properties>
     <maven.compiler.source>1.8</maven.compiler.source>
     <maven.compiler.target>1.8</maven.compiler.target>
