+++ tests/pom.xml
@@ -4,15 +4,15 @@
     <modelVersion>4.0.0</modelVersion>
 
     <parent>
-        <groupId>io.openshift.booster</groupId>
-        <artifactId>spring-boot-cache-parent</artifactId>
+        <groupId>dev.snowdrop.example</groupId>
+        <artifactId>cache</artifactId>
         <version>2.1.3-3-SNAPSHOT</version>
     </parent>
 
     <artifactId>tests</artifactId>
 
-    <name>Spring Boot - Caching Booster - Tests</name>
-    <description>Spring Boot - Caching Booster - Tests</description>
+    <name>Spring Boot - Caching Example - Tests</name>
+    <description>Spring Boot - Caching Example - Tests</description>
 
     <dependencies>
         <dependency>
@@ -127,12 +127,12 @@
                                resources for both modules (which is what the aggregate profile above does)
                             -->
                             <dependency>
-                                <groupId>io.openshift.booster</groupId>
+                                <groupId>dev.snowdrop.example</groupId>
                                 <artifactId>spring-boot-cache-greeting</artifactId>
                                 <version>${project.version}</version>
                             </dependency>
                             <dependency>
-                                <groupId>io.openshift.booster</groupId>
+                                <groupId>dev.snowdrop.example</groupId>
                                 <artifactId>spring-boot-cache-cutename</artifactId>
                                 <version>${project.version}</version>
                             </dependency>
